requirements.txt
appnope==0.1.0
asgiref==3.3.1
astroid==2.4.2
attrs==19.3.0
Automat==0.8.0
autopep8==1.5
azure-common==1.1.23
azure-core==1.5.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.1.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.9
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.3
django-admin-autocomplete-filter==0.5
django-admin-json-editor==0.2.0
django-cacheops==4.2
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-cron==0.5.1
django-dynamic-preferences==1.9
django-filter==2.2.0
django-health-check==3.12.1
django-hint==0.1.5
django-ipware==2.1.0
django-rest-framework==0.1.0
django-summernote==0.8.11.6
django-user-agents==0.4.0
djangorestframework==3.12.2
entrypoints==0.3
Faker==4.14.2
funcy==1.14
httplib2==0.14.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==16.10.1
isort==4.3.21
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
persisting-theory==0.2.1
pika==1.1.0
Pillow==8.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2020.4
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
split-settings==1.0.0
sql-server.pyodbc==1.0
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
text-unidecode==1.3
toml==0.10.1
Twisted==18.9.0
typed-ast==1.4.1
ua-parser==0.10.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
user-agents==2.2.0
validate-docbr==1.8.2
wadllib==1.3.3
wrapt==1.12.1
zipp==1.0.0
zope.interface==4.7.1
